# customers-api/openapi.yaml
openapi: 3.1.0
info:
  title: Customers API
  description: API for managing B2B customers
  version: 1.0.0
  contact:
    name: Andres Buelvas
    email: contact@example.com
servers:
  - url: http://localhost:3001
    description: Local development server
...
paths:
...
  /internal/customers/{id}:
    get:
      tags:
        - Customers
      summary: Get customer by ID (internal use)
      description: Requires service token authentication
      security:
        - ServiceToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '401':
          description: Service token required
        '403':
          description: Invalid service token
        '404':
          description: Customer not found
components:
  schemas:
    Customer:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Acme Corporation
        email:
          type: string
          format: email
          example: contact@acme.com
        phone:
          type: string
          example: '+1-555-0100'
        created_at:
          type: string
          format: date-time
...
  securitySchemes:
    ServiceToken:
      type: http
      scheme: bearer
      description: Service token for internal API communication